Tweak "Best Replacement for Javascript"
Rename it to "A Better Javascript" because JS people tend to be territorial and push back if you try to replace Javascript, but are fine extending it syntactically and semantically in ad-hoc manners. Hopefully saying "This is like Javascript but better" will induce less such push back.

Also change the prose to be more relevant to JS programmers. Nobody knows wtf a "higher-order programming language" means :D :D :D